moved optimizedResized to a module
